T10-13: fix bug from Sabina error handling code.
by Duong Nguyen

+++ server/graphql/resolvers/MongoDB/FavoritesDB.js
@@ -7,18 +7,36 @@ const resolvers = {
         async getAllFavorites(){
             return await FavoriteDBSchema.find({})
         },
-
-        async getFavoriteByUsername(parent, args) {
-            return await FavoriteDBSchema.findOne({username: args.username});
-        },
-
         async deleteFavoriteById(parent, args){
+            const exist =  await FavoriteDBSchema.findOne({username: args._id});
+            if(exist == null){
+                return {
+                    __typename: "IDNotFound",
+                    message: `diese User mit ${args._id} existiert nicht`
+                }
+            }
             return await FavoriteDBSchema.findOneAndDelete({_id: args._id});
         },
-        async deleteFavoriteByUsername(parent, args){
+        async deleteFavotiteByUsername(parent, args){
+            const exist =  await FavoriteDBSchema.findOne({username: args.username});
+            if(exist == null){
+                return {
+                    __typename: "UserNotFound",
+                    message: `${args.username} existiert nicht`
+                }
+            }
             return await FavoriteDBSchema.findOneAndDelete({username:args.username});
         },
-
+        async getFavoriteByUsername(parent, args) {
+            const exist =  await FavoriteDBSchema.findOne({username: args.username});
+            if(exist == null){
+                return {
+                    __typename: "UserNotFound",
+                    message: `${args.username} existiert nicht`
+                }
+            }
+            return await FavoriteDBSchema.findOne({username: args.username});
+        },
     }
 }
 module.exports=resolvers;
